+++ day_6/README.md
@@ -1,1 +1,6 @@
 # Day 6
+
+### Header file
+
+関数の宣言や `#define` とかだけを集めたファイル．
+また，ヘッダファイル内にメモリアドレスを定義しておけば，あとから変更したくなったときにヘッダファイルだけを書き換えれば済む．
