+++ day_6/bootpack.c
@@ -1,59 +1,6 @@
 // コンパイル時に -nostdlib を付与しているので，読み込めない．
 // #include <stdio.h>
-
-#define COL8_000000     0
-#define COL8_FF0000     1
-#define COL8_00FF00     2
-#define COL8_FFFF00     3
-#define COL8_0000FF     4
-#define COL8_FF00FF     5
-#define COL8_00FFFF     6
-#define COL8_FFFFFF     7
-#define COL8_C6C6C6     8
-#define COL8_840000     9
-#define COL8_008400     10
-#define COL8_848400     11
-#define COL8_000084     12
-#define COL8_840084     13
-#define COL8_008484     14
-#define COL8_848484     15
-
-/* Call functions from external files. */
-extern void io_hlt(void);
-extern void io_cli(void);
-extern void re_sprintf(char *str, char *fmt, ...);
-
-/* Call functions from internal files. */
-void init_palette(void);
-void init_screen(char *vram, int x, int y);
-void init_mouse_cursor(char *mouse, char bc);
-void init_gdtidt(void);
-
-void boxfill8(unsigned char *vram, int xsize, unsigned char c, int x0, int y0, int x1, int y1);
-void putfonts8(char *vram, int xsize, int x, int y, char color, char *font);
-void putfonts8_asc(char *vram, int xsize, int x, int y, char color, unsigned char *s);
-void putblock8_8(char *vram, int vxsize, int pxsize, int pysize, int px0, int py0, char *buf, int bxsize);
-
-struct BOOTINFO {
-    char cyls, leds, vmode, reserve;
-    short scrnx, scrny;
-    char *vram;
-};
-
-struct SEGMENT_DESCRIPTOR {
-    short limit_low, base_low;
-    char base_mid, access_right;
-    char limit_high, base_high;
-};
-
-struct GATE_DESCRIPTOR {
-    short offset_low, selector;
-    char dw_count, access_right;
-    short offset_high;
-};
-
-void set_segmdesc(struct SEGMENT_DESCRIPTOR *sd, unsigned int limit, int base, int ar);
-void set_gatedesc(struct GATE_DESCRIPTOR *gd, int offset, int selector, int ar);
+#include "bootpack.h"
 
 void HariMain(void)
 {
