+++ day_6/dsctbl.c
@@ -1,23 +1,5 @@
 // GDT, IDT などの descriptor table 関係
-
-struct SEGMENT_DESCRIPTOR {
-    short limit_low, base_low;
-    char base_mid, access_right;
-    char limit_high, base_high;
-};
-
-struct GATE_DESCRIPTOR {
-    short offset_low, selector;
-    char dw_count, access_right;
-    short offset_high;
-};
-
-void init_gdtidt(void);
-void set_segmdesc(struct SEGMENT_DESCRIPTOR *sd, unsigned int limit, int base, int ar);
-void set_gatedesc(struct GATE_DESCRIPTOR *gd, int offset, int selector, int ar);
-
-extern void load_gdtr(int limit, int addr);
-extern void load_idtr(int limit, int addr);
+#include "bootpack.h"
 
 void init_gdtidt(void)
 {
