+++ day_6/graphic.c
@@ -1,37 +1,5 @@
 // 描画関係
-
-#define COL8_000000     0
-#define COL8_FF0000     1
-#define COL8_00FF00     2
-#define COL8_FFFF00     3
-#define COL8_0000FF     4
-#define COL8_FF00FF     5
-#define COL8_00FFFF     6
-#define COL8_FFFFFF     7
-#define COL8_C6C6C6     8
-#define COL8_840000     9
-#define COL8_008400     10
-#define COL8_848400     11
-#define COL8_000084     12
-#define COL8_840084     13
-#define COL8_008484     14
-#define COL8_848484     15
-
-extern void io_hlt(void);
-extern void io_cli(void);
-extern void io_out8(int port, int data);
-extern int io_load_eflags(void);
-extern void io_store_eflags(int eflags);
-
-void init_palette(void);
-void init_screen(char *vram, int x, int y);
-void init_mouse_cursor(char *mouse, char bc);
-void set_palette(int start, int end, unsigned char *rgb);
-
-void boxfill8(unsigned char *vram, int xsize, unsigned char c, int x0, int y0, int x1, int y1);
-void putfonts8(char *vram, int xsize, int x, int y, char color, char *font);
-void putfonts8_asc(char *vram, int xsize, int x, int y, char color, unsigned char *s);
-void putblock8_8(char *vram, int vxsize, int pxsize, int pysize, int px0, int py0, char *buf, int bxsize);
+#include "bootpack.h"
 
 void init_palette(void)
 {
